Add select op and better long support

+++ compiler/src/main/java/cc/quarkus/qcc/graph2/Test.java
@@ -4,8 +4,8 @@ package cc.quarkus.qcc.graph2;
  *
  */
 public class Test {
-    public int foo(int blah) {
-        int foo = blah;
+    public long foo(long blah) {
+        long foo = blah;
         while (foo < 100) {
             if (blah < 10) {
                 return 14 + blah;
